@@ -1,9 +1,15 @@
 // Native
-import { join } from 'path';
+import { join } from "path";
 
 // Packages
-import { BrowserWindow, app, ipcMain, IpcMainEvent, nativeTheme } from 'electron';
-import isDev from 'electron-is-dev';
+import {
+  BrowserWindow,
+  app,
+  ipcMain,
+  IpcMainEvent,
+  nativeTheme,
+} from "electron";
+import isDev from "electron-is-dev";
 
 const height = 600;
 const width = 800;
@@ -19,12 +25,14 @@ function createWindow() {
     resizable: true,
     fullscreenable: true,
     webPreferences: {
-      preload: join(__dirname, 'preload.js')
-    }
+      preload: join(__dirname, "preload.js"),
+    },
   });
 
   const port = process.env.PORT || 3000;
-  const url = isDev ? `http://localhost:${port}` : join(__dirname, '../dist-vite/index.html');
+  const url = isDev
+    ? `http://localhost:${port}`
+    : join(__dirname, "../dist-vite/index.html");
 
   // and load the index.html of the app.
   if (isDev) {
@@ -33,24 +41,24 @@ function createWindow() {
     window?.loadFile(url);
   }
   // Open the DevTools.
-  // window.webContents.openDevTools();
+  window.webContents.openDevTools();
 
   // For AppBar
-  ipcMain.on('minimize', () => {
+  ipcMain.on("minimize", () => {
     // eslint-disable-next-line no-unused-expressions
     window.isMinimized() ? window.restore() : window.minimize();
     // or alternatively: win.isVisible() ? win.hide() : win.show()
   });
-  ipcMain.on('maximize', () => {
+  ipcMain.on("maximize", () => {
     // eslint-disable-next-line no-unused-expressions
     window.isMaximized() ? window.restore() : window.maximize();
   });
 
-  ipcMain.on('close', () => {
+  ipcMain.on("close", () => {
     window.close();
   });
 
-  nativeTheme.themeSource = 'dark';
+  nativeTheme.themeSource = "dark";
 }
 
 // This method will be called when Electron has finished
@@ -59,7 +67,7 @@ function createWindow() {
 app.whenReady().then(() => {
   createWindow();
 
-  app.on('activate', () => {
+  app.on("activate", () => {
     // On macOS it's common to re-create a window in the app when the
     // dock icon is clicked and there are no other windows open.
     if (BrowserWindow.getAllWindows().length === 0) createWindow();
@@ -69,15 +77,15 @@ app.whenReady().then(() => {
 // Quit when all windows are closed, except on macOS. There, it's common
 // for applications and their menu bar to stay active until the user quits
 // explicitly with Cmd + Q.
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') app.quit();
+app.on("window-all-closed", () => {
+  if (process.platform !== "darwin") app.quit();
 });
 
 // In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and require them here.
 
 // listen the channel `message` and resend the received message to the renderer process
-ipcMain.on('message', (event: IpcMainEvent, message: any) => {
+ipcMain.on("message", (event: IpcMainEvent, message: any) => {
   console.log(message);
-  setTimeout(() => event.sender.send('message', 'common.hiElectron'), 500);
+  setTimeout(() => event.sender.send("message", "common.hiElectron"), 500);
 });
