@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import './style/switch_dark_mode.css';
+import React, { useState } from "react";
+import "./style/switch_dark_mode.css";
 
 const svgMoon = new URL(`./assets/moon.svg`, import.meta.url).href;
 const svgSun = new URL(`./assets/sun.svg`, import.meta.url).href;
@@ -9,7 +9,7 @@ function SwitchDarkMode() {
 
   const darkModeHandler = () => {
     setDark(!isDark);
-    document.body.classList.toggle('dark');
+    document.body.classList.toggle("dark");
   };
 
   return (
@@ -21,7 +21,13 @@ function SwitchDarkMode() {
       <span className="sun">
         <img src={svgSun} alt="sun" />
       </span>
-      <input id="checkbox" type="checkbox" className="input" checked={isDark} onChange={darkModeHandler} />
+      <input
+        id="checkbox"
+        type="checkbox"
+        className="input"
+        checked={isDark}
+        onChange={darkModeHandler}
+      />
       <span className="slider" />
     </label>
   );
