@@ -1,16 +1,17 @@
 const en = {
   translations: {
     common: {
-      selectLanguage: 'Select Language',
-      clickMe: 'Click me',
+      selectLanguage: "Select Language",
+      clickMe: "Click me",
       helloElectron: "Hello I'm from React World",
-      helloBrowser: 'You are in a Browser, so no Electron functions are available',
-      welcome: '💝 Welcome 💝, now send a message to the Main 📩📩',
-      send: 'Send',
-      messageSent: 'Message sent!!',
-      hiElectron: 'hi from electron'
-    }
-  }
+      helloBrowser:
+        "You are in a Browser, so no Electron functions are available",
+      welcome: "💝 Welcome 💝, now send a message to the Main 📩📩",
+      send: "Send",
+      messageSent: "Message sent!!",
+      hiElectron: "hi from electron",
+    },
+  },
 };
 
 export default en;
