@@ -1,16 +1,17 @@
 const es = {
   translations: {
     common: {
-      selectLanguage: 'Seleccione el idioma',
-      clickMe: 'Haga clic aquí',
-      helloElectron: 'Hola, soy de React World',
-      helloBrowser: 'Estás en un navegador, por lo que no hay funciones de Electron disponibles',
-      welcome: '💝 Bienvenido 💝, ahora envía un mensaje al Principal 📩📩',
-      send: 'Enviar',
-      messageSent: '¡¡Mensaje enviado!!',
-      hiElectron: 'hola desde el electrón'
-    }
-  }
+      selectLanguage: "Seleccione el idioma",
+      clickMe: "Haga clic aquí",
+      helloElectron: "Hola, soy de React World",
+      helloBrowser:
+        "Estás en un navegador, por lo que no hay funciones de Electron disponibles",
+      welcome: "💝 Bienvenido 💝, ahora envía un mensaje al Principal 📩📩",
+      send: "Enviar",
+      messageSent: "¡¡Mensaje enviado!!",
+      hiElectron: "hola desde el electrón",
+    },
+  },
 };
 
 export default es;
