@@ -1,7 +1,7 @@
-import enTranslation from './en';
-import esTranslation from './es';
+import enTranslation from "./en";
+import esTranslation from "./es";
 
 export default {
   en: enTranslation,
-  es: esTranslation
+  es: esTranslation,
 };
