@@ -1,22 +1,22 @@
 /** @type {import('tailwindcss').Config} */
-import withMT from '@material-tailwind/react/utils/withMT';
+import withMT from "@material-tailwind/react/utils/withMT";
 
 export default withMT({
-  content: ['./src/index.html', './src/**/*.{js,ts,jsx,tsx}'],
-  darkMode: 'class',
+  content: ["./src/index.html", "./src/**/*.{js,ts,jsx,tsx}"],
+  darkMode: "class",
   theme: {
     extend: {
       colors: {
-        'slate-800': 'rgb(30 41 59)',
-        'slate-900': 'rgb(15 23 42)'
-      }
-    }
+        "slate-800": "rgb(30 41 59)",
+        "slate-900": "rgb(15 23 42)",
+      },
+    },
   },
   variants: {
     extend: {},
     fontFamily: {
-      sans: ['Inter', 'ui-sans-serif', 'system-ui']
-    }
+      sans: ["Inter", "ui-sans-serif", "system-ui"],
+    },
   },
-  plugins: []
+  plugins: [],
 });
