@@ -1,22 +1,22 @@
-import react from '@vitejs/plugin-react';
-import { UserConfig, ConfigEnv } from 'vite';
-import { rmSync } from 'node:fs';
-import { join } from 'path';
-import electron from 'vite-plugin-electron';
-import renderer from 'vite-plugin-electron-renderer';
-import pkg from './package.json';
+import react from "@vitejs/plugin-react";
+import { UserConfig, ConfigEnv } from "vite";
+import { rmSync } from "node:fs";
+import { join } from "path";
+import electron from "vite-plugin-electron";
+import renderer from "vite-plugin-electron-renderer";
+import pkg from "./package.json";
 
 const root = join(__dirname);
-const srcRoot = join(__dirname, 'src');
-rmSync('dist-electron', { recursive: true, force: true });
+const srcRoot = join(__dirname, "src");
+rmSync("dist-electron", { recursive: true, force: true });
 
 const buildElectron = (isDev: boolean) => ({
   sourcemap: isDev,
   minify: !isDev,
-  outDir: join(root, 'dist-electron'),
+  outDir: join(root, "dist-electron"),
   rollupOptions: {
-    external: Object.keys(pkg.dependencies || {})
-  }
+    external: Object.keys(pkg.dependencies || {}),
+  },
 });
 
 function plugins(isDev: boolean) {
@@ -25,76 +25,76 @@ function plugins(isDev: boolean) {
     electron([
       {
         // Main-Process entry file of the Electron App.
-        entry: join(root, 'electron/index.ts'),
+        entry: join(root, "electron/index.ts"),
         onstart(options) {
           options.startup();
         },
         vite: {
-          build: buildElectron(isDev)
-        }
+          build: buildElectron(isDev),
+        },
       },
       {
-        entry: join(root, 'electron/preload.ts'),
+        entry: join(root, "electron/preload.ts"),
         onstart(options) {
           // Notify the Renderer-Process to reload the page when the Preload-Scripts build is complete,
           // instead of restarting the entire Electron App.
           options.reload();
         },
         vite: {
-          build: buildElectron(isDev)
-        }
-      }
+          build: buildElectron(isDev),
+        },
+      },
     ]),
 
-    renderer()
+    renderer(),
   ];
 }
 
 export default ({ command }: ConfigEnv): UserConfig => {
   // DEV
-  if (command === 'serve') {
+  if (command === "serve") {
     return {
       root: srcRoot,
-      base: '/',
+      base: "/",
       plugins: plugins(true),
       resolve: {
         alias: {
-          '/@': srcRoot
-        }
+          "/@": srcRoot,
+        },
       },
       build: {
-        outDir: join(root, '/dist-vite'),
+        outDir: join(root, "/dist-vite"),
         emptyOutDir: true,
-        rollupOptions: {}
+        rollupOptions: {},
       },
       server: {
-        port: process.env.PORT === undefined ? 3000 : +process.env.PORT
+        port: process.env.PORT === undefined ? 3000 : +process.env.PORT,
       },
       optimizeDeps: {
-        exclude: ['path']
-      }
+        exclude: ["path"],
+      },
     };
   }
   // PROD
   return {
     root: srcRoot,
-    base: './',
+    base: "./",
     plugins: plugins(false),
     resolve: {
       alias: {
-        '/@': srcRoot
-      }
+        "/@": srcRoot,
+      },
     },
     build: {
-      outDir: join(root, '/dist-vite'),
+      outDir: join(root, "/dist-vite"),
       emptyOutDir: true,
-      rollupOptions: {}
+      rollupOptions: {},
     },
     server: {
-      port: process.env.PORT === undefined ? 3000 : +process.env.PORT
+      port: process.env.PORT === undefined ? 3000 : +process.env.PORT,
     },
     optimizeDeps: {
-      exclude: ['path']
-    }
+      exclude: ["path"],
+    },
   };
 };
